feat: fetch coingecko
