+++ Cargo.toml
@@ -72,6 +72,7 @@ num-derive = "0.4.2"
 num-traits = "0.2.19"
 proc-macro2 = "1.0.86"
 quote = "1.0.36"
+reqwest = "0.11.27"
 serde = { version = "^1.0", features = ["derive"] }
 serde_json = "1.0.102"
 serde_with = "3.9.0"
