+++ api/Cargo.toml
@@ -21,6 +21,7 @@ http = { workspace = true }
 jito-bytemuck = { workspace = true }
 jito-vault-client = { workspace = true, features = ["serde"] }
 jito-vault-core = { workspace = true }
+reqwest = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 solana-account-decoder = { workspace = true }
