+++ api/src/lib.rs
@@ -1,6 +1,6 @@
 use error::JitoRestakingApiError;
-use serde::{Deserialize, Serialize};
-use solana_program::pubkey::Pubkey;
+// use serde::{Deserialize, Serialize};
+// use solana_program::pubkey::Pubkey;
 
 pub mod error;
 pub mod router;
