+++ api/src/router/vaults/mod.rs
@@ -1,3 +1,2 @@
-pub(crate) mod get_vault;
-pub(crate) mod list_vaults;
 pub(crate) mod tvl;
+pub(crate) mod vault;
