+++ api/src/router/vaults/vault.rs
@@ -1,7 +1,11 @@
-use std::sync::Arc;
+use std::{str::FromStr, sync::Arc};
 
-use anchor_lang::AnchorDeserialize;
-use axum::{extract::State, response::IntoResponse, Json};
+use anchor_lang::{prelude::Pubkey, AnchorDeserialize};
+use axum::{
+    extract::{Path, State},
+    response::IntoResponse,
+    Json,
+};
 use jito_bytemuck::Discriminator;
 use jito_vault_client::{accounts::Vault, programs::JITO_VAULT_ID};
 use solana_account_decoder::UiAccountEncoding;
@@ -10,20 +14,8 @@ use solana_rpc_client_api::{
     filter::{Memcmp, MemcmpEncodedBytes, RpcFilterType},
 };
 
-use crate::router::RouterState;
+use crate::{error::JitoRestakingApiError, router::RouterState};
 
-
-/// Retrieves the history of a specific validator, based on the provided vote account and optional epoch filter.
-///
-/// # Returns
-/// - `Ok(Json(history))`: A JSON response containing the validator history information. If the epoch filter is provided, it only returns the history for the specified epoch.
-///
-/// # Example
-/// This endpoint can be used to fetch the history of a validator's performance over time, either for a specific epoch or for all recorded epochs:
-/// ```
-/// GET /validator_history/{vote_account}?epoch=200
-/// ```
-/// This request retrieves the history for the specified vote account, filtered by epoch 200.
 pub(crate) async fn list_vaults(
     State(state): State<Arc<RouterState>>,
 ) -> crate::Result<impl IntoResponse> {
@@ -54,4 +46,18 @@ pub(crate) async fn list_vaults(
     }
 
     Ok(Json(vaults))
+}
+
+pub(crate) async fn get_vault(
+    State(state): State<Arc<RouterState>>,
+    Path(vault_pubkey): Path<String>,
+) -> crate::Result<impl IntoResponse> {
+    let vault_pubkey = Pubkey::from_str(&vault_pubkey)?;
+    let account = state.rpc_client.get_account(&vault_pubkey).await?;
+    let vault = Vault::deserialize(&mut account.data.as_slice()).map_err(|e| {
+        tracing::warn!("error deserializing Vault: {:?}", e);
+        JitoRestakingApiError::AnchorError(e.into())
+    })?;
+
+    Ok(Json(vault))
 }
